MNT: move widgets/page/ to pages/; import from pages/__init__.py
