+++ squirrel/pages/__init__.py
@@ -1,0 +1,10 @@
+from .collection_builder import CollectionBuilderPage  # noqa
+from .diff import DiffPage  # noqa
+from .entry import (BaseParameterPage, CollectionPage, ParameterPage,  # noqa
+                    ReadbackPage, SetpointPage, SnapshotPage)
+from .page import Page  # noqa
+from .pv_browser import PVBrowserPage  # noqa
+from .search import SearchPage  # noqa
+from .snapshot_comparison import SnapshotComparisonPage  # noqa
+from .snapshot_details import SnapshotDetailsPage  # noqa
+from .tag import TagPage, TagsDialog  # noqa
