+++ squirrel/pages/pv_browser.py
@@ -6,9 +6,9 @@ from qtpy import QtCore, QtGui, QtWidgets
 
 from squirrel.client import Client
 from squirrel.model import PV
+from squirrel.pages import Page
 from squirrel.permission_manager import PermissionManager
 from squirrel.utils import parse_csv_to_dict
-from squirrel.widgets.page.page import Page
 from squirrel.widgets.tables import (PV_BROWSER_HEADER, CSVTableModel,
                                      PVBrowserFilterProxyModel,
                                      PVBrowserTableModel, SquirrelTableView)
