+++ squirrel/pages/snapshot_comparison.py
@@ -3,7 +3,7 @@ from qtpy import QtCore, QtWidgets
 
 from squirrel.client import Client
 from squirrel.model import Snapshot
-from squirrel.widgets.page.page import Page
+from squirrel.pages import Page
 from squirrel.widgets.tables import (COMPARE_HEADER,
                                      SnapshotComparisonTableModel,
                                      SquirrelTableView)
