+++ squirrel/pages/snapshot_details.py
@@ -5,7 +5,7 @@ from qtpy import QtCore, QtWidgets
 
 from squirrel.client import Client
 from squirrel.model import Snapshot
-from squirrel.widgets.page.page import Page
+from squirrel.pages import Page
 from squirrel.widgets.tables import (PV_HEADER, PVTableModel,
                                      SnapshotTableModel, SquirrelTableView)
