+++ squirrel/pages/ui/collection_builder_page.ui
