+++ squirrel/pages/ui/diff_page.ui
