+++ squirrel/pages/ui/name_desc_tags_widget.ui
