+++ squirrel/pages/ui/search_page.ui
